requirements.txt
django==3.0.7
psycopg2==2.8.6
djangorestframework==3.10
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.0
djoser
django-cors-headers
Pillow